数値以外であればクエリを実行しないように修正。

+++ kantan1/kantan_delete.php
@@ -10,7 +10,11 @@
 require_once("data/db_info.php");
 $s = new PDO("mysql:host=$SERV;dbname=$DBNM;charset=utf8", $USER, $PASS);
 $b1_d = $_POST["b1"];
-$s->query("DELETE FROM tbk WHERE bang=$b1_d");
+if (preg_match("/[^0-9]/", $b1_d)) {
+  print "<div style='color:red'>数字以外は入力しないで！！</div>";
+} else {
+  $s->query("DELETE FROM tbk WHERE bang=$b1_d");
+}
 $re = $s->query("SELECT * FROM tbk ORDER BY bang");
 while ($kekka = $re->fetch()) {
   print $kekka[0];
